+++ lib/utils/app_utils.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:notekar/models/palette.dart';
 
-const appVersion = '5.1.5';
-const appBuildNumber = '25';
+const appVersion = '5.1.6';
+const appBuildNumber = '26';
 const appBuildDate = '2026-07-29';
 
 // 8pt Grid Constants